+++ frontend/src/components/MainGame/MainGame.cy.jsx
